little extra fix

+++ dae/oprofiled.c
@@ -256,7 +256,7 @@ static int opd_need_backup_samples_files(void)
 			if (header.ctr_event != ctr_event[header.ctr] ||
 			    header.ctr_um != ctr_um[header.ctr] ||
 			    header.ctr_count != ctr_count[header.ctr] ||
-			    header.cpu_type != cpu_type ||
+			    header.cpu_type != (u32)cpu_type ||
 			    header.separate_samples != separate_samples) {
 				need_backup = 1;
 			}
